feat: extract ThemeProvider
