+++ src/components/index.tsx
@@ -1,3 +1,4 @@
 export * from '@mantine/core'
 export * from './SortableTh'
 export * from './CodeEditor'
+export * from './InseriThemeProvider'
